+++ fortytwo-2fa.php
@@ -5,7 +5,7 @@ Plugin Name: Fortytwo - Two-Factor Authentication plugin
 Plugin URI: https://www.fortytwo.com
 Description: Implement the fortytwo Two Factor Authentication service for login and register on wordpress.
 Version: 1.0.8
-Author: Sebastien Lemarinel <sebastien.lemarinel@fortytwo.com>
+Author: Fortytwo <devs@fortytwo.com>
 Author URI: https://www.fortytwo.com
 License: MIT
 */