+++ src/Artefact/ApiSettingsSection.php
@@ -1,10 +1,12 @@
 <?php
 namespace Fortytwo\Wordpress\Plugin\TwoFactorAuthentication\Artefact;
 
+use Fortytwo\Wordpress\Plugin\TwoFactorAuthentication\Artefact\ArtefactAbstract;
 use Fortytwo\Wordpress\Plugin\TwoFactorAuthentication\Interfaces\SectionInterface;
 use Fortytwo\Wordpress\Plugin\TwoFactorAuthentication\Value\APICodeLengthValue;
 use Fortytwo\Wordpress\Plugin\TwoFactorAuthentication\Value\APICodeTypeValue;
 use Fortytwo\Wordpress\Plugin\TwoFactorAuthentication\Value\APICustomSenderIDValue;
+use Fortytwo\Wordpress\Plugin\TwoFactorAuthentication\Value\APITemplateMessageValue;
 use Fortytwo\Wordpress\Plugin\TwoFactorAuthentication\Value\APICodeCaseSensitiveValue;
 use Fortytwo\Wordpress\Plugin\TwoFactorAuthentication\Value\APICallBackUrlValue;
 
@@ -13,7 +15,7 @@ use Fortytwo\Wordpress\Plugin\TwoFactorAuthentication\Value\APICallBackUrlValue;
  *
  * @license https://opensource.org/licenses/MIT MIT
  */
-class ApiSettingsSection implements SectionInterface
+class ApiSettingsSection extends ArtefactAbstract implements SectionInterface
 {
     /**
      * @inheritDoc
@@ -69,6 +71,15 @@ class ApiSettingsSection implements SectionInterface
             'fortytwo-2fa-admin',
             'ApiSettingsSection'
         );
+
+        add_settings_field(
+            'apiTemplateMessage',
+            'Message Template:',
+            array($this, 'apiTemplateMessageCallback'),
+            'fortytwo-2fa-admin',
+            'ApiSettingsSection'
+        );
+
     }
 
     /**
@@ -84,18 +95,7 @@ class ApiSettingsSection implements SectionInterface
      */
     public function apiCodeLengthCallback()
     {
-        $codeLength = new APICodeLengthValue();
-
-        $html = '<select id="apiCodeLength" name="fortytwo2fa[apiCodeLength]" >';
-        foreach ($codeLength->getOptions() as $value) {
-            if ($codeLength == $value) {
-                $html.= '<option selected="selecte" value="' . $value . '" >' . $value . '</option>';
-            } else {
-                $html.= '<option value="' . $value . '" >' . $value . '</option>';
-            }
-        }
-        $html.= '</select>';
-        echo $html;
+        echo $this->select(new APICodeLengthValue());
     }
 
     /**
@@ -103,19 +103,7 @@ class ApiSettingsSection implements SectionInterface
      */
     public function apiCodeTypeCallback()
     {
-        $codeType = new APICodeTypeValue();
-
-        $html = '<select id="apiCodeType" name="fortytwo2fa[apiCodeType]" >';
-
-        foreach ($codeType->getOptions() as $value) {
-            if ((string)$codeType == $value) {
-                $html.= '<option selected="selecte" value="' . $value . '" >' . $value . '</option>';
-            } else {
-                $html.= '<option value="' . $value . '" >' . $value . '</option>';
-            }
-        }
-        $html.= '</select>';
-        echo $html;
+        echo $this->select(new APICodeTypeValue());
     }
 
     /**
@@ -123,18 +111,7 @@ class ApiSettingsSection implements SectionInterface
      */
     public function apiCaseSensitiveCallback()
     {
-        $codeCaseSensitive = new APICodeCaseSensitiveValue();
-
-        $html = '<select id="apiCaseSensitive" name="fortytwo2fa[apiCaseSensitive]" >';
-        if ((string)$codeCaseSensitive == 'true') {
-            $html.= '<option value="true" selected="selected" >Yes</option>';
-            $html.= '<option value="false" >No</option>';
-        } else {
-            $html.= '<option value="true" >Yes</option>';
-            $html.= '<option value="false" selected="selected" >No</option>';
-        }
-        $html.= '</select>';
-        echo $html;
+        echo $this->select(new APICodeCaseSensitiveValue());
     }
 
     /**
@@ -162,6 +139,21 @@ class ApiSettingsSection implements SectionInterface
             <br>
             <small>Only Alphanumeric and numeric are accepted. Max characters for numeric is 15 & alphanumeric is 11.</small>',
             (string)$senderID
+        );
+    }
+
+    /**
+     * API Message Template.
+     */
+    public function apiTemplateMessageCallback()
+    {
+        $messageTemplate= new APITemplateMessageValue();
+
+        printf(
+            '<input type="text" id="apiTemplateMessage" name="fortytwo2fa[apiTemplateMessage]" value="%s" />
+            <br>
+            <small>The text that should appear in the message the client receives.It is mandatory that ‘{#TFA_CODE}’ is included in this string.</small>',
+            (string)$messageTemplate
         );
     }
 }