+++ src/Controller/Login.php
@@ -7,7 +7,9 @@ use Fortytwo\Wordpress\Plugin\TwoFactorAuthentication\Utils\TemplateEngine;
 use Fortytwo\Wordpress\Plugin\TwoFactorAuthentication\Utils\TrustedDevice;
 use Fortytwo\SDK\TwoFactorAuthentication\TwoFactorAuthentication;
 use Fortytwo\Wordpress\Plugin\TwoFactorAuthentication\Value\TrustedStateValue;
+use Fortytwo\Wordpress\Plugin\TwoFactorAuthentication\Value\LoginMandatoryValue;
 use Fortytwo\Wordpress\Plugin\TwoFactorAuthentication\Value\LoginResendStateValue;
+use Fortytwo\Wordpress\Plugin\TwoFactorAuthentication\Value\RegisterMandatoryValue;
 
 /**
  * Set Login 2fa step UI and add logic
@@ -21,6 +23,7 @@ class Login extends AbstractAuth
         add_action('wp_login', array($this, 'wpLogin'), 10, 2);
         add_action('login_form_validate_2fa', array($this, 'loginFormValidate2fa'), 10, 2);
         add_action('login_form_resend_2fa', array($this, 'resendSMS'), 10, 2);
+        add_action('login_form_add_phone_2fa', array($this, 'savePhone'), 10, 2);
         add_action('login_enqueue_scripts', array($this, 'loginEnqueueScript'), 1);
     }
 
@@ -54,6 +57,33 @@ class Login extends AbstractAuth
         }
     }
 
+    public function savePhone()
+    {
+        $user = get_userdata(intval($_POST['wp-auth-id']));
+
+        if ($user) {
+
+            $phoneValue = esc_attr(get_user_option('2faPhone', $user->ID));
+            wp_clear_auth_cookie();
+
+            if (!$user) {
+                $user = wp_get_current_user();
+            }
+
+            $nonce = new Nonce();
+            $loginNonce = $nonce->create($user->ID);
+
+            if (!$loginNonce) {
+                wp_die(esc_html__('Could not save login nonce.'));
+            }
+
+            $redirectTo = isset($_REQUEST['redirect_to']) ? $_REQUEST['redirect_to'] : $_SERVER['REQUEST_URI'];
+
+            self::addPhoneHtml($user, $loginNonce['key'], $redirectTo);
+            exit;
+        }
+    }
+
 
     /**
      * Display the login form.
@@ -69,32 +99,109 @@ class Login extends AbstractAuth
                 if ((count(array_intersect($options['twoFactorByRole'], $user->roles)) > 0) ||
                     (!isset($options['twoFactorByRole']))
                 ) {
-
                     $phoneValue = esc_attr(get_user_option('2faPhone', $user->ID));
+                    wp_clear_auth_cookie();
+
+                    if (!$user) {
+                        $user = wp_get_current_user();
+                    }
+
+                    $nonce = new Nonce();
+                    $loginNonce = $nonce->create($user->ID);
+
+                    if (!$loginNonce) {
+                        wp_die(esc_html__('Could not save login nonce.'));
+                    }
+
+                    $redirectTo = isset($_REQUEST['redirect_to']) ? $_REQUEST['redirect_to'] : $_SERVER['REQUEST_URI'];
+
                     if (($phoneValue) && ($phoneValue != '')) {
-                        wp_clear_auth_cookie();
-
-                        if (!$user) {
-                            $user = wp_get_current_user();
-                        }
-
-                        $nonce = new Nonce();
-                        $loginNonce = $nonce->create($user->ID);
-
-                        if (!$loginNonce) {
-                            wp_die(esc_html__('Could not save login nonce.'));
-                        }
-
-                        $redirectTo = isset($_REQUEST['redirect_to']) ? $_REQUEST['redirect_to'] : $_SERVER['REQUEST_URI'];
-
                         self::loginHtml($user, $loginNonce['key'], $redirectTo);
                         exit;
+                    } else {
+                        $loginMandatoryPhone = new LoginMandatoryValue();
+                        if ($loginMandatoryPhone->isMandatory()) {
+                            self::addPhoneHtml($user, $loginNonce['key'], $redirectTo);
+                            exit;
+                        }
                     }
                 }
             }
         }
+    }
 
+    /**
+     * Generates the html form for the second step of the authentication process.
+     *
+     * @param object        $user WP_User object of the logged-in user.
+     * @param string        $loginNonce A string nonce stored in usermeta.
+     * @param string        $redirectTo The URL to which the user would like to be redirected.
+     * @param string        $errorMsg Optional. Login error message.
+     * @param string        $clientRef Optional. Client reference to use after error
+     */
+    public function addPhoneHtml($user, $loginNonce, $redirectTo, $errorMsg = false, $clientRef = false)
+    {
+        if ($_REQUEST['2faPhone']) {
+            $this->savePhoneNumber($user->ID);
 
+            self::loginHtml($user, $loginNonce, $redirectTo, $errorMsg = false, $clientRef = false);
+        } else {
+            $options = get_option('fortytwo2fa');
+
+            $interimLogin = false; //isset($_REQUEST['interim-login']);
+
+            // Save rememberme option
+            $rememberme = 0;
+            if (isset($_REQUEST['rememberme']) && $_REQUEST['rememberme']) {
+                $rememberme = 1;
+            }
+
+            //Include the login header
+            login_header();
+
+            //Show error message
+            if (!empty($errorMsg)) {
+                echo '<div id="login_error"><strong>' . esc_html($errorMsg) . '</strong><br /></div>';
+            }
+
+            // Set form action url
+            $formAction = esc_url(
+                set_url_scheme(
+                    add_query_arg(
+                        'action',
+                        'add_phone_2fa',
+                        wp_login_url()
+                    ),
+                    'login_post'
+                )
+            );
+
+            //Hack for capturing the footer
+            ob_start();
+            do_action('login_footer');
+            $wpFooter = ob_get_contents();
+            ob_end_clean();
+
+            // Show template
+            echo TemplateEngine::render(
+                'AddPhoneForm.html',
+                array(
+                    'formAction'    => $formAction,
+                    'userId'        => esc_attr($user->ID),
+                    'authNonce'     => esc_attr($loginNonce),
+                    'interimLogin'  => $interimLogin,
+                    'redirectTo'    => $redirectTo,
+                    'rememberMe'    => esc_attr($rememberme),
+                    'homeUrl'       => esc_url(home_url('/')),
+                    'HomeUrlLabel'  => esc_html(sprintf(__('&larr; Back to %s'), get_bloginfo('title', 'display'))),
+                    'wpFooter'      => $wpFooter,
+                    'clientRef'     => $clientRef,
+                    'resendSMS'     => $resendSMSLoginSection,
+                    'digits'        => $digits,
+                    'trustedDevice' => $trustedDeviceSection
+                )
+            );
+        }
     }
 
     /**
@@ -202,7 +309,7 @@ class Login extends AbstractAuth
             $digits = 4;
         }
 
-        //Hack for capturing the footer
+        // Show template
         echo TemplateEngine::render(
             'CodeValidationForm.html',
             array(