+++ src/Value/AbstractValue.php
@@ -33,4 +33,34 @@ abstract class AbstractValue
     {
         return $this->value;
     }
+
+    /**
+     * Return the value of the object.
+     *
+     * @return mixed value of the object
+     */
+    public function getValue()
+    {
+        return $this->value;
+    }
+
+    /**
+     * Return the field name.
+     *
+     * @return string Field name.
+     */
+    public function getFieldName()
+    {
+        return $this->fieldName;
+    }
+
+    /**
+     * Return the field id.
+     *
+     * @return string Field id.
+     */
+    public function getFieldId()
+    {
+        return $this->fieldId;
+    }
 }