+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 CHANGELOG
 =========
 
+## Version 1.1.0
+_2016-10-04_
+- **[IMPROVEMENT]** Support for Wordpress 4.6.X.
+- **[IMPROVEMENT]** Update SDK dependencies.
+- **[IMPROVEMENT]** Phone field rendering updated.
+- **[IMPROVEMENT]** New API Paremeter : Message template to personalize the message sent with the 2FA Code.
+- **[IMPROVEMENT]** New Mandatory option allowing to have 2FA option as optional or mandatory on login and/or on register.
+
 ## Version 1.0.8
 _2016-05-02_
 - **[DOCUMENTATION]** Fix typo in the readme file.